Move tutorial code to intro package
