+++ samples/tutorial/src/jvmMain/kotlin/com/slack/circuit/tutorial/main.kt
@@ -4,5 +4,5 @@ package com.slack.circuit.tutorial
 
 fun main() {
   // TODO replace with your own impl if following the tutorial!
-  com.slack.circuit.tutorial.impl.main()
+  com.slack.circuit.tutorial.intro.main()
 }
